# api/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateTimeEntryRequest:
    properties:
      description:
        type: string
      duration_minutes:
        minimum: 1
        type: integer
    required:
    - duration_minutes
    type: object
...
  dto.StartTimerRequest:
    properties:
      description:
        maxLength: 5000
        type: string
    type: object
...
  dto.TimeEntryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      ended_at:
        type: string
      id:
        type: string
      started_at:
        type: string
      task_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateTimeEntryRequest:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
    type: object
...
  transport.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
info:
  contact: {}
  description: Self-hosted project management tool API
  title: Plume API
  version: "1.0"
paths:
...
  /projects/{id}/tasks/{taskId}/time-entries:
    get:
      description: Returns all time entries for a task
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeEntryResponse'
            type: array
      summary: List time entries
      tags:
      - time_entries
    post:
      consumes:
      - application/json
      description: Creates a completed time entry with a specified duration
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Time entry details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTimeEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.TimeEntryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Create a manual time entry
      tags:
      - time_entries
  /projects/{id}/tasks/{taskId}/time-entries/{entryId}:
    delete:
      description: Permanently deletes a time entry
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Time Entry ID
        in: path
        name: entryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Delete a time entry
      tags:
      - time_entries
    put:
      consumes:
      - application/json
      description: Updates a time entry's description and/or duration
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Time Entry ID
        in: path
        name: entryId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTimeEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeEntryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Update a time entry
      tags:
      - time_entries
  /projects/{id}/tasks/{taskId}/time-entries/start:
    post:
      consumes:
      - application/json
      description: Starts a running time entry for the current user on a task
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Optional description
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.StartTimerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeEntryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Start a timer
      tags:
      - time_entries
  /projects/{id}/tasks/{taskId}/time-entries/stop:
    post:
      description: Stops the current user's running timer on a task
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeEntryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Stop the active timer
      tags:
      - time_entries
...
swagger: "2.0"